Move serialization.rs
